% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{bob_ross}
\alias{bob_ross}
\title{Joy of Painting Objects}
\format{A data frame with 403 rows and 56 variables}
\usage{
data(bob_ross)
}
\description{
A dataset containing an indicator of which episodes contained various tagged 
objects.  Originally used by fivethirtyeight.com and analyzed 
https://fivethirtyeight.com/features/a-statistical-analysis-of-the-work-of-bob-ross
}
\details{
\itemize{ 
  \item Season. The season number
  \item Episode. The episode number
  \item Aired. Sate episode aired
  \item Title. The title of the episode
  \item aurora_borealis. An indicator tag of this event
  \item barn. An indicator tag of this event
  \item beach. An indicator tag of this event
  \item boat. An indicator tag of this event
  \item bridge. An indicator tag of this event
  \item building. An indicator tag of this event
  \item bushes. An indicator tag of this event
  \item cabin. An indicator tag of this event
  \item cactus. An indicator tag of this event
  \item cirrus. An indicator tag of this event
  \item cliff. An indicator tag of this event
  \item clouds. An indicator tag of this event
  \item conifer. An indicator tag of this event
  \item cumulus. An indicator tag of this event
  \item deciduous. An indicator tag of this event
  \item diane_andre. An indicator tag of this event
  \item dock. An indicator tag of this event
  \item farm. An indicator tag of this event
  \item fence. An indicator tag of this event
  \item fire. An indicator tag of this event
  \item flowers. An indicator tag of this event
  \item fog. An indicator tag of this event
  \item framed. An indicator tag of this event
  \item grass. An indicator tag of this event
  \item guest. An indicator tag of this event
  \item hills. An indicator tag of this event
  \item lake. An indicator tag of this event
  \item lakes. An indicator tag of this event
  \item lighthouse. An indicator tag of this event
  \item mill. An indicator tag of this event
  \item moon. An indicator tag of this event
  \item mountain. An indicator tag of this event
  \item mountains. An indicator tag of this event
  \item night. An indicator tag of this event
  \item ocean. An indicator tag of this event
  \item palm_trees. An indicator tag of this event
  \item path. An indicator tag of this event
  \item person. An indicator tag of this event
  \item portrait. An indicator tag of this event
  \item river. An indicator tag of this event
  \item rocks. An indicator tag of this event
  \item snow. An indicator tag of this event
  \item snowy_mountain. An indicator tag of this event
  \item steve_ross. An indicator tag of this event
  \item structure. An indicator tag of this event
  \item sun. An indicator tag of this event
  \item tree. An indicator tag of this event
  \item trees. An indicator tag of this event
  \item waterfall. An indicator tag of this event
  \item waves. An indicator tag of this event
  \item windmill. An indicator tag of this event
  \item winter. An indicator tag of this event
}
}
\references{
https://github.com/fivethirtyeight/data/tree/master/bob-ross
}
\author{
Walt Hickey
}
\keyword{datasets}
