% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{bp_cyl_8}
\alias{bp_cyl_8}
\title{Row Bind Practice: 8 Cylinders}
\format{A data frame with 14 rows and 9 variables}
\usage{
data(bp_cyl_8)
}
\description{
A dataset containing a cars data set that can be used for practicing row
binding.  A resplit of the \code{mtcars} data set by cylinders with select 
columns removed.
}
\details{
\itemize{ 
  \item car. The name of the car
  \item mpg. Miles per gallon
  \item disp. Displacement (cu.in.)
  \item hp. Horsepower
  \item drat. Rear axle ratio
  \item wt. Weight (1000 lbs)
  \item am. Transmission (0 = automatic, 1 = manual)
  \item gear. Number of forward gears
  \item carb. Number of carburetors
}
}
\keyword{datasets}
