% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{breed_name}
\alias{breed_name}
\title{Cross Tab of Top Breeds and Names from nyc_dogs Data Set}
\format{A data frame with 10 rows and 14 variables}
\usage{
data(breed_name)
}
\description{
A dataset containing a cross tab of breeds and dog names.
}
\details{
\itemize{ 
  \item breed. The breed of the dog
  \item Bella. A name
  \item Buddy. A name 
  \item Charlie. A name 
  \item Coco. A name 
  \item Daisy. A name 
  \item Lola. A name 
  \item Lucky. A name 
  \item Lucy. A name 
  \item Max. A name 
  \item Molly. A name 
  \item Princess. A name 
  \item Rocky. A name 
  \item n. Total count of that breed for the \code{nyc_dogs} data set. 
}
}
\references{
https://fusiontables.google.com/data?docid=1pKcxc8kzJbBVzLu_kgzoAMzqYhZyUhtScXjB0BQ#rows:id=1
}
\keyword{datasets}
