% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{buffalo_snow_tidy}
\alias{buffalo_snow_tidy}
\title{Tidy Version of Historic Snowfall by Month for Buffalo}
\format{A data frame with 932 rows and 4 variables}
\usage{
data(buffalo_snow_tidy)
}
\description{
A dataset containing historical monthly snowfall records.
}
\details{
\itemize{ 
  \item Season. The snow season (spans across 2 years)
  \item Year. The year within the season
  \item Month. The month within the year
  \item Decade. The decade attached to the year
  \item Snow. The amount of snow in inches that fell
}
}
\references{
https://www.weather.gov/buf/BuffaloSnow
}
\keyword{datasets}
