% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{cereal}
\alias{cereal}
\title{Ceareal Nutrition}
\format{A data frame with 77 rows and 15 variables}
\usage{
data(cereal)
}
\description{
A dataset containing nutritional information on popular cereals.
}
\details{
\itemize{ 
  \item name. Name of the cereal
  \item mfr. Manufacturer of cereal \code{list(A = 'American Home Food Products', 
  G = 'General Mills', K = 'Kelloggs', N = 'Nabisco', P = 'Post', Q = 'Quaker Oats',
  R = 'Ralston Purina')}
  \item type. Cereal type; hot (h) or cold (c)
  \item calories. Calories per serving
  \item protein. Grams of protei
  \item fat. Grams of fat
  \item sodium. Milligrams of sodium 
  \item fiber. Grams of dietary fiber
  \item carbo. Grams of complex carbohydrates
  \item sugars. Grams of sugars
  \item potass. Milligrams of potassium
  \item vitamins. Vitamins and minerals: One of 0, 25, or 100, indicating the typical percentage of FDA recommended
  \item shelf. Display shelf (1, 2, or 3, counting from the floor)
  \item weight. Weight in ounces of one serving
  \item cups. Number of cups in one serving
}
}
\references{
https://www.kaggle.com/crawford/80-cereals
}
\author{
Petra Isenberg, Pierre Dragicevic and Yvonne Jansen
}
\keyword{datasets}
