% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{package}
\name{exampledata}
\alias{exampledata}
\alias{package-exampledata}
\alias{exampledata-package}
\title{Example Data for Teaching}
\description{
Small data sets that are optimized for teaching particular constructs.
}
