% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{fake_sales_person}
\alias{fake_sales_person}
\title{A Fake Salesperson Data Set}
\format{A data frame with 10 rows and 2 variables}
\usage{
data(fake_sales_person)
}
\description{
A dataset containing fake sales people that can be matched to the 
\code{carnegie} data set for practice joining 
(\code{by = c("REGIONID" = "REGIONID")}).
}
\details{
\itemize{ 
  \item PERSON. A salesperson
  \item REGIONID. Region code
}
}
\keyword{datasets}
