% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{full_carnegie_dat}
\alias{full_carnegie_dat}
\title{The Full, Raw Carnegie Data Set}
\format{A data frame with 4,665 rows and 94 variables}
\usage{
data(full_carnegie_dat)
}
\description{
A dataset containing all columns from the 2015 Carnegie data set.  This data 
is raw, meaning it will likely need to be cleaned before it is usable.
}
\details{
See http://carnegieclassifications.iu.edu/downloads/CCIHE2015-PublicDataFile.xlsx
for the original data set from Carnegie.  In order to understand the variables
you will need the 'Variables' sheet/tab in the .xlsx file.  In order to recode
the categorical variables you will need \code{dplyr::case_when} and the 'Labels'
sheet/tabl in the .xlsx file.
}
\references{
http://carnegieclassifications.iu.edu/
}
\keyword{datasets}
