% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{full_college_scorecard_dat}
\alias{full_college_scorecard_dat}
\title{The Full, Raw College Score Card Data Set}
\format{A data frame with 7,593 rows and 1,825 variables}
\usage{
data(full_college_scorecard_dat)
}
\description{
The Full, Raw College Score Card Data Set
}
\details{
A dataset containing all columns from the 2015 College Score Card data set.  
This data is raw, meaning it will likely need to be cleaned before it is usable.
Cleaning may mean recoding variables and/or coercing data from one class type 
to another (see the examples section for an example coercing from character to
numeric).  See https://collegescorecard.ed.gov/data/documentation/ for 
documentation of the variable meanings.
}
\examples{
\dontrun{
library(dplyr)

full_college_scorecard_dat \%>\%
    select(TUITIONFEE_IN, PFTFTUG1_EF) \%>\%
    mutate_all(as.numeric) \%>\%
    ggplot(aes(TUITIONFEE_IN, PFTFTUG1_EF)) +
        geom_jitter()
}
}
\references{
https://collegescorecard.ed.gov/data/documentation/
}
\keyword{datasets}
