% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{jp_health}
\alias{jp_health}
\title{Join Practice: Sex Based Health Info}
\format{A data frame with 2 rows and 3 variables}
\usage{
data(jp_health)
}
\description{
A dataset containing the sex based health information for practice with
joins.  Tables that work together for join practice include: \code{jp_health},
\code{jp_publishers}, & \code{superheroes}.
}
\details{
\itemize{ 
  \item sex. male/female
  \item life_expectancy. Average life expectancy in 2016 in the U.S.
  \item daily_calories. Recommended daily caloric intake
}
}
\keyword{datasets}
