% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{jp_publishers}
\alias{jp_publishers}
\title{Join Practice: Publisher Info}
\format{A data frame with 3 rows and 2 variables}
\usage{
data(jp_publishers)
}
\description{
A dataset containing superhero comic book publisher founding years for 
practice with joins.  Tables that work together for join practice include: 
\code{jp_health}, \code{jp_publishers}, & \code{superheroes}.
}
\details{
\itemize{ 
  \item publisher. The name of the comic book publisher
  \item yr_founded. Year the publishing company was founded
}
}
\references{
http://stat545.com/bit001_dplyr-cheatsheet.html
}
\author{
Jenny Bryan
}
\keyword{datasets}
