% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{jp_superheroes}
\alias{jp_superheroes}
\title{Join Practice:}
\format{A data frame with 7 rows and 4 variables}
\usage{
data(jp_superheroes)
}
\description{
A dataset containing the superhero character information for practice with
joins.  Tables that work together for join practice include: \code{jp_health},
\code{jp_publishers}, & \code{superheroes}.
}
\details{
\itemize{ 
  \item name. The name of the character
  \item alignment. Does the character  align with good or bad more often?
  \item sex. The sex of the character
  \item publisher. The publisher who owns the rights to the character
}
}
\references{
http://stat545.com/bit001_dplyr-cheatsheet.html
}
\author{
Jenny Bryan
}
\keyword{datasets}
