% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{nyc_dogs}
\alias{nyc_dogs}
\title{Dogs of NYC Project}
\format{A data frame with 81,542 rows and 15 variables}
\usage{
data(nyc_dogs)
}
\description{
A dataset containing New York City dog data.
}
\details{
\itemize{ 
  \item dog_name. Name of dog
  \item gender. Gender of dog
  \item breed. Breed of dog
  \item birth. Birth date of dog
  \item birth_year. Birth year of dog
  \item birth_month. Birth month of dog
  \item dominant_color. Primary fur color
  \item secondary_color. Secondary fur color
  \item third_color. Third fur color
  \item spayed_or_neutered. Logical indicating if dog has been fixed
  \item guard_or_trained. Logical indicating if the dog is trained for gaurd duty or or special area
  \item borough. Borough of NYC dog is located in
  \item zip_code. Zip code of NYC dog is located in
  \item latitude. Latitude of NYC dog is located in
  \item longitude. Longitude of NYC dog is located in
}
}
\references{
https://fusiontables.google.com/data?docid=1pKcxc8kzJbBVzLu_kgzoAMzqYhZyUhtScXjB0BQ#rows:id=1
}
\keyword{datasets}
