% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{traffic_violations}
\alias{traffic_violations}
\title{Montgomery County Traffic Violations}
\format{A data frame with 1,276,580 rows and 35 variables}
\usage{
data(traffic_violations)
}
\description{
A dataset containing traffic violation records from January 1, 2015 to 
December 31, 2017.  Taken from: https://data.montgomerycountymd.gov/Public-Safety/Traffic-Violations/4mse-ku6q
First read about this data set from: https://www.nytimes.com/2018/01/30/upshot/do-fast-and-furious-movies-cause-a-rise-in-speeding.html
}
\details{
\itemize{ 
  \item `Date Of Stop`. Date of the traffic violation.
  \item `Time Of Stop`. Time of the traffic violation.
  \item `Latitude`. Latitude location of the traffic violation.
  \item `Longitude`. Longitude location of the traffic violation.
  \item `Accident`. If traffic violation involved an accident.
  \item `Belts`. If traffic violation involved a seat belt violation.
  \item `Personal Injury`. If traffic violation involved Personal Injury.
  \item `Property Damage`. If traffic violation involved Property Damage.
  \item `Fatal`. If traffic violation involved a fatality.
  \item `Commercial License`. If driver holds a Commercial Drivers License.
  \item `HAZMAT`. If the traffic violation involved hazardous materials.
  \item `Commercial Vehicle`. If the vehicle committing the traffic violation is a commercial vehicle.
  \item `Work Zone`. If the traffic violation was in a work zone.
  \item `State`. State issuing the vehicle registration.
  \item `VehicleType`. Type of vehicle (Examples: Automobile, Station Wagon, Heavy Duty Truck, etc.)
  \item `Year`. Year vehicle was made.
  \item `Make`. Manufacturer of the vehicle (Examples: Ford, Chevy, Honda, Toyota, etc.)
  \item `Model`. Model of the vehicle.
  \item `Color`. Color of the vehicle.
  \item `Violation Type`. Violation type. (Examples: Warning, Citation, SERO)
  \item `Charge`. Numeric code for the specific charge.
  \item `Contributed To Accident`. If the traffic violation was a contributing factor in an accident.
  \item `Race`. Race of the driver. (Example: Asian, Black, White, Other, etc.)
  \item `Gender`. Gender of the driver (F = Female, M = Male)
  \item `Driver City`. City of the driver's home address.
  \item `Driver State`. State of the driver's home address.
  \item `Arrest Type`. Type of Arrest (A = Marked, B = Unmarked, etc.)
}
}
\examples{
\dontrun{
# https://www.nytimes.com/2018/01/30/upshot/do-fast-and-furious-movies-cause-a-rise-in-speeding.html
# https://data.montgomerycountymd.gov/Public-Safety/Traffic-Violations/4mse-ku6q

data(traffic_violations)

library(tidyverse)
library(chron)
library(numform)
library(ggrepel)

traffic_violations_clean <- traffic_violations \%>\%
    mutate(
        `Date Of Stop` = as.Date(`Date Of Stop`, format = '\%m/\%d/\%Y'),
        year_of_stop = format(`Time Of Stop`, format="\%Y"),
        hour_of_day = `Time Of Stop` \%>\%
            strptime(format="\%H:\%M:\%S") \%>\%
            format(format = '\%H') \%>\%
            as.numeric(),
        day_of_week = factor(weekdays(`Date Of Stop`), levels = numform::constant_weekdays)
    ) \%>\%
    mutate_at(
        vars(Accident:`Work Zone`, `Contributed To Accident`), 
        function(x) {
            case_when(
                x == 'No' ~ FALSE,
                x == 'Yes' ~ TRUE,
                TRUE ~ NA
            )
        }
    ) \%>\%
    select(-Geolocation) \%>\%
    extract(VehicleType, c('Vehicle_Type_Code', 'Vehicle_Type'), '(\\\\d+)\\\\s+-\\\\s+([A-M].+)\\\\s*')

traffic_violations_clean \%>\%
    group_by(day_of_week, hour_of_day) \%>\%
    summarize(Count = n()) \%>\%
    ggplot(aes(hour_of_day, Count, group = 1)) +
        geom_line() +
        facet_wrap(~day_of_week, ncol = 2) +
        scale_x_continuous(breaks = seq(0, 23, by = 2), 
            labels = function(x) numform::f_12_hour(x, format = "\%I \%p")
        ) +
        scale_y_continuous(labels = f_denom)



traffic_violations_clean \%>\%
    group_by(day_of_week, hour_of_day) \%>\%
    summarize(Count = n()) \%>\%
    ungroup() \%>\%
    mutate(weekend = day_of_week \%in\% c('Sunday', 'Saturday', 'Friday')) \%>\%
    ggplot(aes(hour_of_day, Count, group = day_of_week, color = day_of_week)) +
        geom_line(aes(linetype = weekend)) +
        scale_x_continuous(
            limits = c(-2, 23),
            breaks = seq(0, 23, by = 2), 
            labels = function(x) numform::f_12_hour(x, format = "\%I \%p")
        ) +
        scale_y_continuous(labels = f_denom) +
        geom_text_repel(
            data = traffic_violations_clean \%>\%
                count(day_of_week, hour_of_day) \%>\%
                rename(Count = n) \%>\%
                filter(hour_of_day == 0),
            aes(label = day_of_week),
            size = 4,
            nudge_x = -1,
            segment.color = 'grey80'
          ) +
          theme(legend.position = 'none')


traffic_violations_clean \%>\%
    group_by(day_of_week, hour_of_day) \%>\%
    summarize(Count = n()) \%>\%
    ggplot(aes(hour_of_day, Count, group = day_of_week, color = day_of_week)) +
        geom_line() +
        scale_x_continuous(breaks = seq(0, 23, by = 2), 
            labels = function(x) numform::f_12_hour(x, format = "\%I \%p")
        ) +
        scale_y_continuous(labels = f_denom) +
        coord_polar()

## Indicates this sample doesn't typically contain DUI/DWI
traffic_violations_clean  \%>\%
    filter(Alcohol) \%>\%
    group_by(day_of_week, hour_of_day) \%>\%
    summarize(Count = n()) \%>\%
    ungroup() \%>\%
    mutate(weekend = day_of_week \%in\% c('Sunday', 'Saturday', 'Friday')) \%>\%
    ggplot(aes(hour_of_day, Count, group = day_of_week, color = day_of_week)) +
        geom_line(aes(linetype = weekend)) +
        scale_x_continuous(
            limits = c(-2, 23),
            breaks = seq(0, 23, by = 2), 
            labels = function(x) numform::f_12_hour(x, format = "\%I \%p")
        ) +
        scale_y_continuous(labels = f_denom) +
        geom_text_repel(
            data = traffic_violations_clean  \%>\%
                filter(Alcohol) \%>\%
                count(day_of_week, hour_of_day) \%>\%
                rename(Count = n) \%>\%
                filter(hour_of_day == 0),
            aes(label = day_of_week),
            size = 4,
            nudge_x = -1,
            segment.color = 'grey80'
          ) +
          theme(legend.position = 'none')
}
}
\references{
https://data.montgomerycountymd.gov/Public-Safety/Traffic-Violations/4mse-ku6q
}
\keyword{datasets}
