% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exampledata-package.R
\docType{data}
\name{wegmans}
\alias{wegmans}
\title{A Fake Wegmans Grocery}
\format{A data frame with 30 rows and 10 variables}
\usage{
data(wegmans)
}
\description{
A dataset containing fake grocery related data.
}
\details{
\itemize{
  \item department. The department
  \item item. The grocery item
  \item last_shipment. The last time the item was shipped to the stor
  \item weight. The weight of the item
  \item wholesale. The price Weagmans pays for the item
  \item old_price. The price Weagmans was charging before the price change
  \item new_price. The current price
  \item popularity. A customer rating of stars (1-5 stars; with 5 being the most popular)
  \item organic. A messy character field describing if the item comes in organic ('yes', 'y'), non-organic ('no', 'n'), or both organic and non-organic
  \item product_of_usa. A dummy coded (aka, one-hot encoding) variable (0-no, 1-yes) of whether the item is from the U.S. or not
}
}
\keyword{datasets}
